requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
authlib
requests-oauthlib
fastapi-sessions
sqlalchemy
pydantic[email]
google-auth
google-auth-oauthlib
google-auth-httplib2
